refactor(calendar): integrate picker components into date-picker
