--- packages/components/calendar/src/date-picker-field.tsx
@@ -1,46 +1,0 @@
-import type { CSSUIObject, HTMLUIProps } from "@yamada-ui/core"
-import { ui, forwardRef } from "@yamada-ui/core"
-import { PopoverTrigger } from "@yamada-ui/popover"
-import { cx } from "@yamada-ui/utils"
-import { useDatePickerContext } from "./use-date-picker"
-
-type DatePickerFieldOptions = {
-  inputProps?: HTMLUIProps<"input">
-}
-
-export type DatePickerFieldProps = HTMLUIProps<"div"> & DatePickerFieldOptions
-
-export const DatePickerField = forwardRef<DatePickerFieldProps, "input">(
-  ({ className, h, minH, inputProps, ...rest }, ref) => {
-    const styles = useDatePickerContext()
-
-    const css: CSSUIObject = {
-      pe: "2rem",
-      h,
-      minH,
-      display: "flex",
-      alignItems: "center",
-      ...styles.field,
-      cursor: "text",
-      ...styles.field,
-    }
-
-    return (
-      <PopoverTrigger>
-        <ui.div
-          className={cx("ui-date-picker__field", className)}
-          __css={css}
-          {...rest}
-        >
-          <ui.input
-            ref={ref}
-            className="ui-date-picker__field__input"
-            display="inline-block"
-            w="full"
-            {...inputProps}
-          />
-        </ui.div>
-      </PopoverTrigger>
-    )
-  },
-)
